chore: use Rust stable on CI
While doing that, the whole CI got a bit of a clean up. E.g. use
stock circle CI docker images instead of custom ones. Also downscale
the resources classes that were used. It now takes less than 25% of
the credits.

Also update the code to fix any warnings caused by using a newer
Rust version.

+++ rust/src/fvm/externs.rs
@@ -44,7 +44,7 @@ impl Rand3 for CgoExterns {
             let mut buf = [0u8; 32];
             match cgo_extern_get_chain_randomness(
                 self.handle,
-                pers as i64,
+                pers,
                 round,
                 entropy.as_ptr(),
                 entropy.len() as i32,
@@ -73,7 +73,7 @@ impl Rand3 for CgoExterns {
             let mut buf = [0u8; 32];
             match cgo_extern_get_beacon_randomness(
                 self.handle,
-                pers as i64,
+                pers,
                 round,
                 entropy.as_ptr(),
                 entropy.len() as i32,
